apps/hello-world/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
kafka-python==2.0.2